+++ app/build.gradle
@@ -63,5 +63,9 @@ dependencies {
 
     //Material Design
     implementation("com.google.android.material:material:1.6.1")
+
+    //Koin
+    implementation "io.insert-koin:koin-android:3.2.0"
+    implementation "io.insert-koin:koin-core:3.2.0"
 }
