+++ app/src/main/java/com/internship/move/MainActivity.kt
@@ -1,20 +1,31 @@
 package com.internship.move
 
-import android.app.Application
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.internship.move.di.internalStorage
+import com.internship.move.di.repositories
+import com.internship.move.di.viewModels
+import org.koin.android.ext.koin.androidContext
+import org.koin.core.context.GlobalContext.startKoin
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        _appContext = this.application
+        initKoin()
     }
 
-    companion object {
-        private var _appContext: Application? = null
-        val appContext: Application
-            get() = _appContext ?: throw IllegalArgumentException("App context not initialized")
+    private fun initKoin() {
+        startKoin {
+            androidContext(this@MainActivity)
+            modules(
+                listOf(
+                    internalStorage,
+                    viewModels,
+                    repositories
+                )
+            )
+        }
     }
 }
