+++ app/src/main/java/com/internship/move/data/model/User.kt
@@ -1,0 +1,7 @@
+package com.internship.move.data.model
+
+data class User(
+    val email: String,
+    val username: String,
+    val password: String
+)
