--- app/src/main/java/com/internship/move/feature/authentification/viewmodel/AuthenticationViewModel.kt
@@ -1,5 +1,0 @@
-package com.internship.move.feature.authentification.viewmodel
-
-import com.internship.move.MainViewModel
-
-class AuthenticationViewModel : MainViewModel()
