--- app/src/main/java/com/internship/move/feature/map/viewmodel/MapViewModel.kt
@@ -1,5 +1,0 @@
-package com.internship.move.feature.map.viewmodel
-
-import com.internship.move.MainViewModel
-
-class MapViewModel : MainViewModel()
