--- app/src/main/java/com/internship/move/feature/onboarding/adapter/OnSkipListener.kt
@@ -1,5 +1,0 @@
-package com.internship.move.feature.onboarding.adapter
-
-interface OnSkipListener {
-    fun onSkipClickListener()
-}
