--- app/src/main/java/com/internship/move/feature/onboarding/adapter/ViewPagerAdapter.kt
@@ -1,12 +1,0 @@
-package com.internship.move.feature.onboarding.adapter
-
-import androidx.fragment.app.Fragment
-import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.internship.move.feature.onboarding.model.ViewPagerItem
-
-class ViewPagerAdapter(private val currentList: ArrayList<ViewPagerItem>, fragment: Fragment) : FragmentStateAdapter(fragment) {
-
-    override fun createFragment(position: Int): Fragment = ViewPagerFragment.newInstance(currentList[position])
-
-    override fun getItemCount(): Int = currentList.size
-}
