--- app/src/main/java/com/internship/move/feature/splash/viewmodel/SplashViewModel.kt
@@ -1,5 +1,0 @@
-package com.internship.move.feature.splash.viewmodel
-
-import com.internship.move.MainViewModel
-
-class SplashViewModel : MainViewModel()
