+++ app/src/main/java/com/internship/move/presentation/authentification/forgotpassword/ForgotPasswordFragment.kt
@@ -1,0 +1,41 @@
+package com.internship.move.presentation.authentification.forgotpassword
+
+import android.os.Bundle
+import android.view.View
+import androidx.core.widget.doOnTextChanged
+import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.findNavController
+import com.internship.move.R
+import com.internship.move.databinding.FragmentForgotPasswordBinding
+import com.zhuinden.fragmentviewbindingdelegatekt.viewBinding
+
+class ForgotPasswordFragment : Fragment(R.layout.fragment_forgot_password) {
+
+    private val binding by viewBinding(FragmentForgotPasswordBinding::bind)
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        initListeners()
+    }
+
+    private fun initListeners() {
+        binding.backIV.setOnClickListener {
+            findNavController().navigate(ForgotPasswordFragmentDirections.actionForgotPasswordFragmentToLoginFragment())
+        }
+
+        binding.sendResetBtn.setOnClickListener {
+            findNavController().navigate(ForgotPasswordFragmentDirections.actionForgotPasswordFragmentToResetPasswordFragment())
+        }
+
+        binding.emailTIET.doOnTextChanged { _, _, _, _ ->
+            changeBtnState(
+                binding.emailTIET.text?.isNotEmpty() ?: false,
+            )
+        }
+    }
+
+    private fun changeBtnState(isNotEmpty: Boolean) {
+        binding.sendResetBtn.isEnabled = isNotEmpty
+    }
+}
