+++ app/src/main/java/com/internship/move/presentation/onboarding/adapter/OnSkipListener.kt
@@ -1,0 +1,5 @@
+package com.internship.move.presentation.onboarding.adapter
+
+interface OnSkipListener {
+    fun onSkipClickListener()
+}
