+++ app/src/main/java/com/internship/move/presentation/onboarding/adapter/ViewPagerFragment.kt
@@ -1,0 +1,41 @@
+package com.internship.move.presentation.onboarding.adapter
+
+import android.os.Bundle
+import android.view.View
+import androidx.core.content.ContextCompat
+import androidx.core.view.isVisible
+import androidx.fragment.app.Fragment
+import com.internship.move.R
+import com.internship.move.data.model.OnboardingViewPagerItem
+import com.internship.move.databinding.ItemOnboardingPageBinding
+import com.zhuinden.fragmentviewbindingdelegatekt.viewBinding
+
+class ViewPagerFragment : Fragment(R.layout.item_onboarding_page) {
+
+    private val binding by viewBinding(ItemOnboardingPageBinding::bind)
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val pagerItem: OnboardingViewPagerItem? = arguments?.getParcelable(KEY_ONBOARDING_PAGER_ITEM)
+
+        if (pagerItem != null) {
+            binding.onboardingIV.setImageDrawable(ContextCompat.getDrawable(requireContext(), pagerItem.photo))
+            binding.onboardingHeadingTV.text = getText(pagerItem.heading)
+            binding.onboardingContentTV.text = getText(pagerItem.content)
+            binding.onboardingSkipTV.isVisible = !pagerItem.isLastItem
+        }
+
+        binding.onboardingSkipTV.setOnClickListener {
+            (parentFragment as? OnSkipListener)?.onSkipClickListener()
+        }
+    }
+
+    companion object {
+        private const val KEY_ONBOARDING_PAGER_ITEM = "KEY_ONBOARDING_PAGER_ITEM"
+
+        fun newInstance(pagerItem: OnboardingViewPagerItem): ViewPagerFragment = ViewPagerFragment().apply {
+            arguments = Bundle().apply { putParcelable(KEY_ONBOARDING_PAGER_ITEM, pagerItem) }
+        }
+    }
+}
