+++ app/src/main/java/com/internship/move/repository/Repository.kt
@@ -1,0 +1,20 @@
+package com.internship.move.repository
+
+import com.internship.move.utils.InternalStorageManager
+
+class Repository(
+    private val internalStorageManager: InternalStorageManager
+) {
+
+    fun getIsUserLoggedIn() = internalStorageManager.getIsUserLoggedIn()
+
+    fun getHasUserCompletedOnboarding() = internalStorageManager.getHasUserCompletedOnboarding()
+
+    fun setHasUserCompletedOnboarding(boolean: Boolean) {
+        internalStorageManager.setUserHasCompletedOnboarding(boolean)
+    }
+
+    fun setIsUserLoggedIn(boolean: Boolean) {
+        internalStorageManager.setIsUserLoggedIn(boolean)
+    }
+}
