+++ app/src/main/java/com/internship/move/utils/InternalStorageManager.kt
@@ -1,0 +1,31 @@
+package com.internship.move.utils
+
+import android.content.Context
+import android.content.SharedPreferences
+import com.internship.move.presentation.authentification.register.RegisterFragment
+import com.internship.move.presentation.map.MapFragment
+
+class InternalStorageManager(
+    context: Context
+) {
+    private val sharedPreferences: SharedPreferences =
+        context.getSharedPreferences(KEY_APP_PREFERENCES, Context.MODE_PRIVATE)
+
+    fun getIsUserLoggedIn(): Boolean = sharedPreferences.getBoolean(MapFragment.KEY_IS_USER_LOGGED_IN, false)
+
+
+    fun getHasUserCompletedOnboarding(): Boolean = sharedPreferences.getBoolean(RegisterFragment.KEY_HAS_USER_COMPLETED_ONBOARDING, false)
+
+
+    fun setIsUserLoggedIn(isUserLoggedIn: Boolean) {
+        sharedPreferences.edit().putBoolean(MapFragment.KEY_IS_USER_LOGGED_IN, isUserLoggedIn).apply()
+    }
+
+    fun setUserHasCompletedOnboarding(hasUserCompletedOnboarding: Boolean) {
+        sharedPreferences.edit().putBoolean(RegisterFragment.KEY_HAS_USER_COMPLETED_ONBOARDING, hasUserCompletedOnboarding).apply()
+    }
+
+    companion object {
+        const val KEY_APP_PREFERENCES = "KEY_APP_PREFERENCES"
+    }
+}
