+++ app/src/main/res/layout/fragment_forgot_password.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/bg_app"
     android:orientation="vertical"
-    tools:context=".feature.authentification.forgotpassword.ForgotPasswordFragment"
+    tools:context=".presentation.authentification.forgotpassword.ForgotPasswordFragment"
     tools:ignore="ContentDescription">
 
     <ImageView
