+++ app/src/main/res/layout/fragment_register.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/bg_app"
     android:orientation="vertical"
-    tools:context=".feature.authentification.register.RegisterFragment"
+    tools:context=".presentation.authentification.register.RegisterFragment"
     tools:ignore="ContentDescription">
 
     <ImageView
