+++ app/src/main/res/layout/fragment_reset_password.xml
@@ -1,49 +1,30 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/background_color"
-    tools:context=".feature.authentification.forgotpassword.ResetPasswordFragment"
+    android:background="@drawable/bg_app"
+    android:orientation="vertical"
+    tools:context=".presentation.authentification.forgotpassword.ResetPasswordFragment"
     tools:ignore="ContentDescription">
 
     <ImageView
-        android:id="@+id/topRectangleIV"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/ic_top_rectangle"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageView
-        android:id="@+id/bottomRectangleIV"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/ic_bottom_rectangle"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <ImageView
-        android:id="@+id/chevronIV"
+        android:id="@+id/backIV"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="@dimen/horizontal_margin"
         android:layout_marginTop="@dimen/authentication_chevron_margin_top"
-        android:src="@drawable/ic_chevron_left"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        android:src="@drawable/ic_arrow_back" />
 
     <TextView
         android:id="@+id/titleTV"
-        style="@style/AuthentificationHeadingTV"
+        style="@style/Heading1"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="@dimen/horizontal_margin"
         android:layout_marginTop="@dimen/authentication_forgot_password_title_margin_top"
-        android:text="@string/authentication_reset_password_text"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/chevronIV" />
+        android:text="@string/authentication_reset_password_text" />
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/newPasswordTIL"
@@ -53,9 +34,7 @@
         android:layout_marginTop="40dp"
         android:hint="@string/authentication_reset_password_new_pw_text_view"
         app:endIconDrawable="@drawable/ic_password_toggle"
-        app:endIconMode="password_toggle"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/titleTV">
+        app:endIconMode="password_toggle">
 
         <com.google.android.material.textfield.TextInputEditText
             android:id="@+id/newPasswordTIET"
@@ -73,10 +52,7 @@
         android:layout_marginTop="26dp"
         android:hint="@string/authentication_reset_password_confirm_pw_text_view"
         app:endIconDrawable="@drawable/ic_password_toggle"
-        app:endIconMode="password_toggle"
-        app:errorTextAppearance="@style/ErrorTextView"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/newPasswordTIL">
+        app:endIconMode="password_toggle">
 
         <com.google.android.material.textfield.TextInputEditText
             android:id="@+id/confirmPasswordTIET"
@@ -88,16 +64,14 @@
 
     <Button
         android:id="@+id/resetPasswordBtn"
-        style="@style/TransparentButton"
+        style="@style/PrimaryButton"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="@dimen/horizontal_margin"
         android:layout_marginTop="50dp"
         android:layout_marginEnd="@dimen/horizontal_margin"
-        android:background="@drawable/bg_state_button"
+        android:background="@drawable/bg_primary_button"
         android:enabled="false"
         android:text="@string/authentication_reset_password_text"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/confirmPasswordTIL" />
-</androidx.constraintlayout.widget.ConstraintLayout>
+        android:textColor="@color/primary_button_text_color" />
+</LinearLayout>
