+++ app/src/main/res/navigation/authentification_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/registerFragment"
-        android:name="com.internship.move.feature.authentification.register.RegisterFragment"
+        android:name="com.internship.move.presentation.authentification.register.RegisterFragment"
         android:label="fragment_register"
         tools:layout="@layout/fragment_register">
 
@@ -34,7 +34,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.internship.move.feature.authentification.login.LoginFragment"
+        android:name="com.internship.move.presentation.authentification.login.LoginFragment"
         android:label="fragment_login"
         tools:layout="@layout/fragment_login">
 
@@ -71,7 +71,7 @@
 
     <fragment
         android:id="@+id/forgotPasswordFragment"
-        android:name="com.internship.move.feature.authentification.forgotpassword.ForgotPasswordFragment"
+        android:name="com.internship.move.presentation.authentification.forgotpassword.ForgotPasswordFragment"
         android:label="fragment_forgot_password"
         tools:layout="@layout/fragment_forgot_password">
 
@@ -97,7 +97,7 @@
 
     <fragment
         android:id="@+id/resetPasswordFragment"
-        android:name="com.internship.move.feature.authentification.forgotpassword.ResetPasswordFragment"
+        android:name="com.internship.move.presentation.authentification.forgotpassword.ResetPasswordFragment"
         android:label="fragment_reset_password"
         tools:layout="@layout/fragment_reset_password">
