+++ app/src/main/res/navigation/home_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/mapFragment"
-        android:name="com.internship.move.feature.map.MapFragment"
+        android:name="com.internship.move.presentation.map.MapFragment"
         android:label="fragment_map"
         tools:layout="@layout/fragment_map">
