+++ app/src/main/res/navigation/onboarding_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/onboardingFragment"
-        android:name="com.internship.move.feature.onboarding.OnboardingFragment"
+        android:name="com.internship.move.presentation.onboarding.OnboardingFragment"
         android:label="OnboardingFragment">
 
         <action
@@ -31,7 +31,7 @@
 
     <fragment
         android:id="@+id/helpFragment"
-        android:name="com.internship.move.feature.help.HelpFragment"
+        android:name="com.internship.move.presentation.help.HelpFragment"
         android:label="fragment_help"
         tools:layout="@layout/fragment_help">
