+++ app/src/main/res/navigation/splash_graph.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/splashFragment"
-        android:name="com.internship.move.feature.splash.SplashFragment"
+        android:name="com.internship.move.presentation.splash.SplashFragment"
         android:label="SplashFragment">
 
         <action
