+++ app/src/main/res/values/strings.xml
@@ -38,5 +38,5 @@
     <string name="authentication_reset_password_text">Reset Password</string>
     <string name="authentication_reset_password_new_pw_text_view">New password</string>
     <string name="authentication_reset_password_confirm_pw_text_view">Confirm new password</string>
-    <string name="authentication_reset_password_error">Passwords do not match.</string>
+    <string name="authentication_reset_password_error">Passwords do not match</string>
 </resources>
