+++ app/src/main/res/values/styles.xml
@@ -53,7 +53,7 @@
         <item name="android:layout_marginStart">@dimen/horizontal_margin</item>
         <item name="android:layout_marginEnd">@dimen/horizontal_margin</item>
         <item name="android:hint">This is a hint</item>
-        <item name="hintTextColor">@color/input_box_hint_color</item>
+        <item name="hintTextColor">@color/authentication_second_text_color</item>
         <item name="endIconMode">custom</item>
         <item name="boxStrokeWidth">2dp</item>
         <item name="boxStrokeColor">@color/input_layout_box_stroke_color</item>
@@ -64,11 +64,10 @@
 
     <style name="PrimaryET" parent="Widget.MaterialComponents.TextInputEditText.FilledBox">
         <item name="android:textSize">@dimen/regular_text_size</item>
-        <item name="android:textColor">@color/authentication_text_color</item>
+        <item name="android:textColor">@color/authentication_second_text_color</item>
         <item name="android:singleLine">true</item>
         <item name="android:background">@android:color/transparent</item>
         <item name="fontFamily">@font/bai_jamjuree_regular</item>
-        <item name="android:layout_marginStart">@dimen/edit_text_margin_start</item>
         <item name="android:textCursorDrawable">@null</item>
         <item name="android:inputType">text|textNoSuggestions</item>
     </style>
